create usePost hook
